+++ public/pages/CreateIndex/components/IndexDetail/IndexDetail.tsx
@@ -523,7 +523,7 @@ const IndexDetail = (
                       fullWidth
                       helpText={
                         <div>
-                          Define how documents and their fields are stored and indexed.
+                          Define how documents and their fields are stored and indexed.{" "}
                           <EuiLink target="_blank" external href="https://opensearch.org/docs/latest/opensearch/mappings/">
                             Learn more.
                           </EuiLink>
