+++ public/pages/CreateIndex/components/MappingLabel/index.ts
@@ -1,0 +1,4 @@
+import MappingLabel from "./MappingLabel";
+
+export default MappingLabel;
+export * from "./MappingLabel";
